+++ R/ranger_classification.R
@@ -1,13 +1,17 @@
 #' Run ranger with parameters of data.frame rows.
 #'
 #' This functions calls ranger using the parameter values in each row of the 
-#' provided master_grid, using the data of the list elements.
+#' provided master_grid, using the data of the list elements. Please have
+#' a look at the [ranger doc](https://cran.r-project.org/web/packages/ranger/ranger.pdf)
+#' for explanation on the ranger related variables, the arguments are beginning 
+#' with "ranger" in the description. Except for `the list`, `master_grid` and `.row`
+#' all arguments need to be column names of `master_grid`
 #'
-#' @param Target The respective column from the master_grid
-#' @param ML_object The respective column from the master_grid
-#' @param Cycle The respective column from the master_grid
-#' @param Number_of_trees The respective column from the master_grid
-#' @param Mtry_factor The respective column from the master_grid
+#' @param Target char, the response variable
+#' @param ML_object factor or char, the name of the corresponding `the_list` item
+#' @param Cycle integer, the current repetition
+#' @param Number_of_trees ranger, integer, number of trees per forest
+#' @param Mtry_factor ranger, factor to multiply default ranger mtry argument
 #' @param .row current row of master_grid
 #' @param the_list The input tables list
 #' @param master_grid the data frame containing all parameter combinations