+++ NEWS.md
@@ -3,6 +3,7 @@ Version 0.5
 *  Vignette added demonstrating a possible workflow using built-in data set
 *  Naming patterns changed to open up for users
 *  oversample function is now more appropriately called augment function
+*  relative abundance is used for filtering
 *  ranger regression mtry set to 1/3 of available independent variables as default
 *  speedyseq support added for tax_glom